% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addrepo.R
\name{add}
\alias{add}
\title{Enable a universe}
\usage{
add(universe = "ropensci")
}
\arguments{
\item{universe}{vector with name(s) of the universe(s), i.e. the subdomain
part of \verb{https://ropensci.r-universe.dev}.}
}
\value{
the updated list of repositories
}
\description{
Add r-universe package repositories to your \code{options("repos")} such
that they are enabled by default in \code{install.packages()}. This is not
permanent, so you need to save this in your \verb{~/.Rprofile} in order to
automatically enable this for every R session.
}
