% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{all_universes}
\alias{all_universes}
\title{All universes}
\usage{
all_universes(type = c("all", "universes", "maintainers"))
}
\arguments{
\item{type}{Type of universe to query: "all",
only organizations ("universe"),
personal accounts ("maintainer").}
}
\value{
A character vector of all universes.
}
\description{
All universes
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
head(all_universes())
\dontshow{\}) # examplesIf}
}
