% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_munger_helper_objects.R
\docType{data}
\name{DataMelter}
\alias{DataMelter}
\title{Handle melting of data according to the taxonomy of the input.}
\format{\preformatted{Class 'R6ClassGenerator' <DataMelter> object generator
  Public:
    melt_data: NA
    initialize: function (scale) 
    melt_df_cartesian: function (bucketed_df, by_col) 
    melt_df_uniform: function (bucketed_df, by_col) 
    melt_df_factor: function (bucketed_df, by_col) 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:universe>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "DataMelter_generator"
}}
\usage{
DataMelter
}
\description{
Handle melting of data according to the taxonomy of the input.
}
\keyword{datasets}

