% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_munger.R
\docType{data}
\name{DataMunger}
\alias{DataMunger}
\title{Handle everything to do with producing a melted dataframe to feed into plot
methods later.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataMunger
}
\description{
Handle everything to do with producing a melted dataframe to feed into plot
methods later.
}
\keyword{datasets}

