% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{everyone}
\alias{everyone}
\title{All universes}
\usage{
everyone(type = c("all", "universes", "maintainers"))
}
\arguments{
\item{type}{Type of universe to query: "all",
only organizations ("universe"),
personal accounts ("maintainer").}
}
\value{
A character vector of all universes.
}
\description{
All universes
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
head(everyone())
\dontshow{\}) # examplesIf}
}
