% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_plotter.R
\docType{data}
\name{GgplotPlotter}
\alias{GgplotPlotter}
\title{ggplot2 plotting backend.
Uses some gtable hacks to get exposure bars on a scaled secondary axis.
Unfortunately, this approach means the plot method can't return a ggplot
object, it just draws the plot to whatever graphics device is available.
Can potentially handle extra ggplot arguments another way? With an
additional_arguments argument? Not sure there's a clean solution.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GgplotPlotter
}
\description{
From: http://rpubs.com/kohske/dual_axis_in_ggplot2
}
\keyword{datasets}

