% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_plotter.R
\docType{data}
\name{PlotlyPlotter}
\alias{PlotlyPlotter}
\title{Create plotly objects from a melted dataframe.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PlotlyPlotter
}
\description{
Create plotly objects from a melted dataframe.
}
\keyword{datasets}

