% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{ru_search}
\alias{ru_search}
\title{Search the R universe}
\usage{
ru_search(q, universe = NULL, ...)
}
\arguments{
\item{q}{string to search for. Wrap in double quotes to get exact matches.}

\item{universe}{which universe to target. Default is all universes.}

\item{...}{extra parameters subject to change.}
}
\value{
dataframe with description data
}
\description{
Free text search through package titles, authors, description, topics, etc.
}
\details{
The parameters and output fields are subject to change.
}
