% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{repo_snapshot}
\alias{repo_snapshot}
\alias{r_version}
\title{Download snapshot}
\usage{
repo_snapshot(repo, destdir = "snapshot", bin_versions = r_version())

r_version()
}
\arguments{
\item{repo}{url of the cran-like repository to snapshot}

\item{bin_versions}{vector with versions of R to download the win/mac binary
packages. The default is to download binaries only for your local version.
Set to NULL to not download any binaries.}
}
\description{
Downloads a full copy (snapshot) of an R package repository.
}
\examples{
repo_snapshot('https://jeroen.r-universe.dev', bin_versions = c("4.1", "4.2", "4.3"))
}
