% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe-package.R
\docType{package}
\name{universe-package}
\alias{universe}
\alias{universe-package}
\title{universe: Tools for Working with R-universe <https://r-universe.dev>}
\description{
Utilities to interact with the R-universe platform. Includes functions to manage local package repositories, as well as API wrappers for retrieving data and metadata about packages in r-universe.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/universe/}
  \item \url{https://github.com/ropensci/universe}
  \item Report bugs at \url{https://github.com/ropensci/universe/issues}
}

}
\author{
\strong{Maintainer}: Jeroen Ooms \email{jeroenooms@gmail.com} (\href{https://orcid.org/0000-0002-4035-0289}{ORCID})

Authors:
\itemize{
  \item Maëlle Salmon \email{maelle.salmon@yahoo.se} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})
}

}
\keyword{internal}
