% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/universe.R
\name{universe}
\alias{universe}
\title{Simple univariate plots.}
\usage{
universe(input_df, plot_cols, by_col, buckets = 10, cut_type = "even",
  scale = "uniform")
}
\arguments{
\item{input_df}{input data.frame or data.table}

\item{plot_cols}{string vector naming one or more columns in input_df to
plot against by_col}

\item{by_col}{string naming the x-variable against which the plot_cols are
measured}

\item{buckets}{number of buckets into which to attempt to reduce the data}

\item{cut_type}{either "even" or "quantile". Defaults to "even".
\itemize{
  \item{"even"}{ Cut data into evenly spaced buckets from the min to the max
    of by_col}
  \item{"quantile"}{ Cut data by quantile, in which each bucket will have
    approximately the same weight.}
}}

\item{scale}{either "uniform" or "cartesian". Defaults to "uniform".
\itemize{
  \item{"uniform"}{ Distribute data buckets evenly. Represent weight with a
    bar chart.}
  \item{"cartesian"}{ Distribute data buckets according to the mean of their
    by_var. Represent weight with a histogram.}
}}
}
\value{
\pkg{plotly} object.
}
\description{
Bucket data and draw univariate plots in plotly. Univariate plots show the
variation of one or more variable versus one covariate. universe will also
draw a neat exposure histogram.
}
\details{
Operates in two parts; bucketing the data into discrete bins (either evenly
in by_col or evenly by quantile in by_col), then drawing the variation of
the plot_cols vs by_col with a histogram of by_col's exposure.
}

