% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{universe_all_packages}
\alias{universe_all_packages}
\title{Info on all packages in an universe}
\usage{
universe_all_packages(universe, limit = 100L)
}
\arguments{
\item{universe}{Name of the universe (character of length 1)}

\item{limit}{Number of results to return (integer of length 1)}
}
\value{
A list with information on all packages in the universe.
}
\description{
Info on all packages in an universe
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
universe_all_packages("jeroen")
universe_all_packages("ropensci")
\dontshow{\}) # examplesIf}
}
\seealso{
Other universe: 
\code{\link{universe_ls}()},
\code{\link{universe_one_package}()},
\code{\link{universe_search}()}
}
\concept{universe}
