% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{universe_ls}
\alias{universe_ls}
\title{All packages in an universe}
\usage{
universe_ls(universe)
}
\arguments{
\item{universe}{Name of the universe (character of length 1)}
}
\value{
A character vector of all packages in the universe.
}
\description{
All packages in an universe
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
universe_ls("jeroen")
universe_ls("ropensci")
\dontshow{\}) # examplesIf}
}
\seealso{
Other universe: 
\code{\link{universe_all_packages}()},
\code{\link{universe_one_package}()},
\code{\link{universe_search}()}
}
\concept{universe}
