% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{universe_one_package}
\alias{universe_one_package}
\title{Info on a single packages in an universe}
\usage{
universe_one_package(universe, package)
}
\arguments{
\item{universe}{Name of the universe (character of length 1)}

\item{package}{Name of the package (character of length 1)}
}
\value{
A list with information on the package.
}
\description{
Info on a single packages in an universe
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
universe_one_package("jeroen", package = "curl")
\dontshow{\}) # examplesIf}
}
\seealso{
Other universe: 
\code{\link{universe_all_packages}()},
\code{\link{universe_ls}()},
\code{\link{universe_search}()}
}
\concept{universe}
