% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{universe_query}
\alias{universe_query}
\title{Query the API of a single universe}
\usage{
universe_query(universe_url, path, query_params = NULL)
}
\arguments{
\item{universe_url}{URL to the universe like ""https://jeroen.r-universe.dev"}

\item{path}{Path to append after api like "ls"}

\item{query_params}{Named list of query parameters}
}
\value{
The JSON from the API as a list
}
\description{
Query the API of a single universe
}
\keyword{internal}
