% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{universe_search}
\alias{universe_search}
\title{Search within a single universe}
\usage{
universe_search(universe, query, limit = 100L)
}
\arguments{
\item{universe}{Name of the universe (character of length 1)}

\item{query}{Query string.
See \href{https://docs.r-universe.dev/browse/search.html}{R-universe docs}.}

\item{limit}{Number of results to return (integer of length 1)}
}
\value{
A list with query results. The \code{total} field indicates the
total number of results and can be used as \code{limit} value in a second call.
}
\description{
Search within a single universe
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
universe_search("ropensci", query = '"weather data"')
universe_search("ropensci", query = 'needs:httr2')
\dontshow{\}) # examplesIf}
}
\seealso{
Other universe: 
\code{\link{universe_all_packages}()},
\code{\link{universe_ls}()},
\code{\link{universe_one_package}()}
}
\concept{universe}
