% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-json.R
\name{use_custom_manifest}
\alias{use_custom_manifest}
\title{Create \code{packages.json} for existing universe}
\usage{
use_custom_manifest(universe, path = "packages.json")
}
\arguments{
\item{universe}{}
}
\value{
The path to the JSON file it created.
}
\description{
Create \code{packages.json} for existing universe
}
\examples{
\dontshow{if (rlang::is_interactive()) withAutoprint(\{ # examplesIf}
json_file <- withr::local_tempfile()
use_custom_manifest("jeroen", json_file)
file.edit(json_file)
\dontshow{\}) # examplesIf}
}
