% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-gha.R
\name{use_universe_action}
\alias{use_universe_action}
\title{Use R-universe GitHub Actions workflow}
\usage{
use_universe_action(universe = NULL, path = ".")
}
\arguments{
\item{universe}{Universe to set the context to, by default the owner of the
repository.
This affects where R package dependencies are downloaded from
(besides the default repositories).}

\item{path}{Path to the local repository for which you want to use the workflow.}
}
\value{
The path to the workflow
}
\description{
Test the R-universe build workflow from your own GitHub repository
}
\examples{
\dontshow{if (rlang::is_interactive()) withAutoprint(\{ # examplesIf}
temp_dir <- withr::local_tempdir()
workflow_path <- use_universe_action("maelle", temp_dir)
file.edit(workflow_path)
\dontshow{\}) # examplesIf}
}
