% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_munger_helper_objects.R
\docType{data}
\name{VectorBucketer}
\alias{VectorBucketer}
\title{Handle bucketing of the by variable.}
\format{\preformatted{Class 'R6ClassGenerator' <VectorBucketer> object generator
  Public:
    cut_fn: NA
    initialize: function (cut_type) 
    cut_vector: function (cut_vector, buckets) 
    cut_evenly: function (cut_vector, buckets) 
    cut_by_quantile: function (cut_vector, buckets) 
    cut_factor_by_exposure: function (cut_vector, buckets) 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:universe>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "VectorBucketer_generator"
}}
\usage{
VectorBucketer
}
\description{
Handle bucketing of the by variable.
}
\keyword{datasets}

