% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addrepo.R
\name{add}
\alias{add}
\title{Enable a package repository from r-universe}
\usage{
add(universe = "ropensci")
}
\arguments{
\item{universe}{vector with name(s) of the universe(s), i.e. the subdomain
part of \verb{https://ropensci.r-universe.dev}.}
}
\value{
the updated list of repositories
}
\description{
Adds r-universe package repositories to your \code{options("repos")} such
that they are used by default in \code{install.packages()}. If the universe
was already enabled, it will not be added again, hence it is harmless
to call this function multiple times.
}
\details{
Note that changes to your options are not permanent. To automatically
enable a repository for every R session, you can call this function in
your \code{\link[=Startup]{~/.Rprofile}} script.
}
