% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_plotter.R
\docType{data}
\name{GgplotPlotter}
\alias{GgplotPlotter}
\title{ggplot2 plotting backend.
Uses some gtable hacks to get exposure bars on a scaled secondary axis.
Unfortunately, this approach means the plot method can't return a ggplot
object, it just draws the plot to whatever graphics device is available.
Can potentially handle extra ggplot arguments another way? With an
additional_arguments argument? Not sure there's a clean solution.}
\format{\preformatted{Class 'R6ClassGenerator' <GgplotPlotter> object generator
  Public:
    plot_fn: NA
    initialize: function (scale, xtitle, ytitle) 
    plot_fn_cartesian: function (data_munger) 
    plot_fn_uniform: function (data_munger) 
    gg_base_lines: function (data_munger) 
    gg_exposure_bars: function (data_munger) 
    gg_exposure_hist: function (data_munger) 
    ggplot_double_axis: function (plot_left, plot_right) 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:universe>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "GgplotPlotter_generator"
}}
\usage{
GgplotPlotter
}
\description{
From: http://rpubs.com/kohske/dual_axis_in_ggplot2
}
\keyword{datasets}

