% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly_plotter.R
\docType{data}
\name{PlotlyPlotter}
\alias{PlotlyPlotter}
\title{Create plotly objects from a melted dataframe.}
\format{\preformatted{Class 'R6ClassGenerator' <PlotlyPlotter> object generator
  Public:
    plot_fn: NA
    xaxis: NA
    yaxis: NA
    yaxis2: NA
    initialize: function (scale, xtitle, ytitle) 
    plot_fn_hist: function (data_munger) 
    plot_fn_bar: function (data_munger) 
    plot_lines: function (data_munger) 
    add_histogram_trace: function (plot, data_munger) 
    add_bar_trace: function (plot, data_munger) 
    set_layout: function (p) 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:universe>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "PlotlyPlotter_generator"
}}
\usage{
PlotlyPlotter
}
\description{
Create plotly objects from a melted dataframe.
}
\keyword{datasets}

