% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{repo_snapshot}
\alias{repo_snapshot}
\alias{r_version}
\title{Download snapshot}
\usage{
repo_snapshot(
  repo,
  destdir = NULL,
  types = c("src", "win", "mac", "linux", "wasm", "docs"),
  bin_versions = r_version(),
  verbose = interactive()
)

r_version()
}
\arguments{
\item{repo}{url of the cran-like repository to snapshot}

\item{destdir}{path to directory where to save the snapshots}

\item{types}{which files to include. Must be subset of "src", "win", "mac", "linux", "wasm", "docs".}

\item{bin_versions}{vector with versions of R to download binary packages.
Set to NULL to download all binaries available.}

\item{verbose}{print some progress}
}
\value{
None, used for its side-effects.
}
\description{
Downloads a full copy (snapshot) of an R package repository.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
repo_snapshot("https://jeroen.r-universe.dev")
\dontshow{\}) # examplesIf}
}
