% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{repo_snapshot}
\alias{repo_snapshot}
\alias{r_version}
\title{Download snapshot}
\usage{
repo_snapshot(
  repo,
  types = NULL,
  destdir = "snapshot",
  bin_versions = r_version()
)

r_version()
}
\arguments{
\item{repo}{url of the cran-like repository to snapshot}

\item{destdir}{where to save the snapshots (path).}

\item{bin_versions}{vector with versions of R to download binary
packages. The default is to download binaries for all available versions.
Set to NULL to not download any binaries.}

\item{type}{type of snapshot. Either \code{NULL} or one of "src", "win", "mac", "linux", "wasm", "docs".
By default, if \code{NULL}, "win", "mac", "linux" snapshots will be downloaded.}
}
\value{
None, used for its side-effects.
}
\description{
Downloads a full copy (snapshot) of an R package repository.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
repo_snapshot(
  "https://jeroen.r-universe.dev",
   bin_versions = c("4.1", "4.2", "4.3")
)
\dontshow{\}) # examplesIf}
}
