% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_bucketer.R
\docType{data}
\name{VectorBucketer}
\alias{VectorBucketer}
\title{Handle bucketing of the by variable.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VectorBucketer
}
\description{
Handle bucketing of the by variable.
}
\keyword{datasets}

